䷪ refactored the code
