+++ frontend/src/apps/main/assets/png/logo.png
