+++ frontend/src/apps/main/redux/constants.ts
@@ -7,5 +7,5 @@ export const actionType = {
   UPDATE_PROJECTS_SCENE: "UPDATE_PROJECTS_SCENE",
   UPDATE_LANDING_SCENE: "UPDATE_LANDING_SCENE",
   UPDATE_SETTINGS: "UPDATE_SETTINGS",
-  UPDATE_LANDING_LAYOUT: "UPDATE_LANDING_LAYOUT",
+  UPDATE_LAYOUT: "UPDATE_LAYOUT",
 };
