--- frontend/src/apps/main/scenes/articles/style.scss
@@ -1,1 +1,0 @@
-@import "../../../../common/style/variables";
