--- frontend/src/apps/main/scenes/learn/style.scss
@@ -1,1 +1,0 @@
-@import "../../../../common/style/variables";
