+++ frontend/src/apps/main/types/index.ts
@@ -1,7 +1,10 @@
-import { LayoutInitialState } from "t9/apps/main/redux/reducers/layout-reducer";
 import { mainStore } from "t9/apps/main/redux";
 
 export type MainStoreStateInterface = typeof mainStore.getState;
+
+const state = mainStore.getState();
+
+export type StateInterface = typeof state;
 
 export interface MainSettings {
   darkMode: boolean;
@@ -12,9 +15,4 @@ export interface SidebarTreeItem {
   content: string;
   link?: string;
   children?: SidebarTreeItem[];
-}
-
-export interface ReduxState {
-  layout: LayoutInitialState;
-  settings: MainSettings;
 }
